add auto ticket buyer
